+++ gauntlet/packages/gauntlet-solana-contracts/src/commands/contracts/ocr2/index.ts
@@ -8,7 +8,6 @@ import PayRemaining from './payRemaining'
 import ReadState from './read'
 import SetBilling from './setBilling'
 import SetConfig from './setConfig'
-import SetConfigDeployer from './setConfig.deployer'
 import SetPayees from './setPayees'
 import SetupFlow from './setup.dev.flow'
 import SetValidatorConfig from './setValidatorConfig'
@@ -17,7 +16,6 @@ import Transmit from './transmit.dev'
 export default [
   Initialize,
   OCR2InitializeFlow,
-  SetConfigDeployer,
   SetBilling,
   PayRemaining,
   SetPayees,
