+++ gauntlet/packages/gauntlet-solana-contracts/src/core/crypto.ts
