+++ gauntlet/packages/gauntlet-solana-contracts/src/core/scalarMult.ts
