+++ gauntlet/packages/gauntlet-solana-contracts/src/core/utils.ts
